refactor(hr): restructure pages into folder pattern with CSS modules
