--- src/pages/hr/Attendance.tsx
@@ -1,9 +1,0 @@
-const Attendance = () => {
-  return (
-    <div>
-      <h1>Attendance</h1>
-    </div>
-  );
-};
-
-export default Attendance;
