+++ src/pages/hr/Attendance/Attendance.module.css
@@ -1,0 +1,3 @@
+.container {
+  padding: 16px;
+}
