+++ src/pages/hr/Attendance/Attendance.tsx
@@ -1,0 +1,16 @@
+import { Card, Typography } from 'antd';
+
+const { Title } = Typography;
+
+export const Attendance: React.FC = () => {
+  return (
+    <div>
+      <Card>
+        <Title level={3}>Attendance</Title>
+        <p>Attendance page - coming soon</p>
+      </Card>
+    </div>
+  );
+};
+
+export default Attendance;
