+++ src/pages/hr/EmployeeDetails/EmployeeDetails.module.css
@@ -1,0 +1,15 @@
+.container {
+  padding: 16px;
+}
+
+.formGrid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 16px;
+}
+
+@media (max-width: 768px) {
+  .formGrid {
+    grid-template-columns: 1fr;
+  }
+}
