+++ src/pages/hr/EmployeesList/EmployeesList.module.css
@@ -1,0 +1,14 @@
+.container {
+  padding: 16px;
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 16px;
+}
+
+.header h1 {
+  margin: 0;
+}
