--- src/pages/hr/LeaveRequests.tsx
@@ -1,9 +1,0 @@
-const LeaveRequests = () => {
-  return (
-    <div>
-      <h1>Leave Requests</h1>
-    </div>
-  );
-};
-
-export default LeaveRequests;
