+++ src/pages/hr/LeaveRequests/LeaveRequests.module.css
@@ -1,0 +1,3 @@
+.container {
+  padding: 16px;
+}
