+++ src/pages/hr/LeaveRequests/LeaveRequests.tsx
@@ -1,0 +1,16 @@
+import { Card, Typography } from 'antd';
+
+const { Title } = Typography;
+
+export const LeaveRequests: React.FC = () => {
+  return (
+    <div>
+      <Card>
+        <Title level={3}>Leave Requests</Title>
+        <p>Leave requests page - coming soon</p>
+      </Card>
+    </div>
+  );
+};
+
+export default LeaveRequests;
